Exchange entropy-mnemonics library for go-bip39
We removed all traces of entropy-mnemonics and adapted the code to
use the go-bip39 library.

We wanted  to update the codification of words for seeds to a more
standard one.

The only remainder of entropy-mnemonics are in the API, as some URL
still accept the dictionary parameter, but it is ignored by the application