Remove references to Sia in documentation
We have renamed all references in documentation from sia to rivine,
from siacoin to coin, and from siafund to blockstake.

Although there are still references in the documentation most of the
time when is code naming fields or functions that contains the sia word.

Also we have removed documentation on rivinec missing features.

Al other has been renamed or removed.