ensure block structure remains unchanged
fixes #305

While transactions are allowed to be in new/alternative versions,
the block structure should remain as it was,
given that there is no sane way to expect any other way.

MinerPayouts were already defined as a slice of CoinOutputs,
but as the CoinOutput was changed, it modified also the Block structure because of this,
therefore we now define a new structure MinerPayout,
which defines a structure exactly as the legacy CoinOutput was,
and which can remain to be used to the Block structure's purpses.