remove go-tip tests
Even though the Tip (master) version of Golang is pretty stable most of the time,
it does every few months break something in any project it seems.

This is to be expected, as it is the wild west of master in the end,
however it makes us lose time when it happens.
In fact it makes us lose time all the time, as the test stages for Go Tip
take about 250% the time of versioned Go tests.