make the UnlockCondition code more flexible with Go types
this commit relates to issue #404

The problem before this commit was that when a chain implemented
a custom version of the MultiSigCondition, that it would no longer be functional
within the Rivine codebase, which is used as a library for that chain.
Meaning that the wallet module for that chain wouldn't be able to handle its MultiSigConditions.

This commit does not resolve #404, but it may serve as a starting point.